+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Make NGINX optional on KVM, by ignoring existing NGINX IC App CRs which were managed by cluster-operator.
+
 ## [0.23.13] - 2020-07-28
 
 ### Changed